[fix]avoid too much memory in join probe operator

+++ be/src/pipeline/exec/operator.cpp
@@ -243,7 +243,7 @@ void PipelineXLocalStateBase::clear_origin_block() {
 }
 
 Status PipelineXLocalStateBase::filter_block(const vectorized::VExprContextSPtrs& expr_contexts,
-                                             vectorized::Block* block, int column_to_keep) {
+                                             vectorized::Block* block, size_t column_to_keep) {
     RETURN_IF_ERROR(vectorized::VExprContext::filter_block(expr_contexts, block, column_to_keep));
 
     _estimate_memory_usage += vectorized::VExprContext::get_memory_usage(expr_contexts);
